requirements.txt
nose
python-docx
zipfile

